+++ tag-20180828/alibabacloud_tag20180828/client.py
@@ -386,6 +386,7 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> tag_20180828_models.CreatePolicyResponse:
         """
+        ###
         This topic provides an example on how to call the API operation to create a tag policy named `test`. In this example, the Tag Policy feature in multi-account mode is used. The tag policy defines that resources to which the `CostCenter:Beijing` or `CostCenter:Shanghai` tag is added are compliant and other resources are not compliant.
         
         @param request: CreatePolicyRequest
@@ -437,6 +438,7 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> tag_20180828_models.CreatePolicyResponse:
         """
+        ###
         This topic provides an example on how to call the API operation to create a tag policy named `test`. In this example, the Tag Policy feature in multi-account mode is used. The tag policy defines that resources to which the `CostCenter:Beijing` or `CostCenter:Shanghai` tag is added are compliant and other resources are not compliant.
         
         @param request: CreatePolicyRequest
@@ -487,6 +489,7 @@ class Client(OpenApiClient):
         request: tag_20180828_models.CreatePolicyRequest,
     ) -> tag_20180828_models.CreatePolicyResponse:
         """
+        ###
         This topic provides an example on how to call the API operation to create a tag policy named `test`. In this example, the Tag Policy feature in multi-account mode is used. The tag policy defines that resources to which the `CostCenter:Beijing` or `CostCenter:Shanghai` tag is added are compliant and other resources are not compliant.
         
         @param request: CreatePolicyRequest
@@ -500,6 +503,7 @@ class Client(OpenApiClient):
         request: tag_20180828_models.CreatePolicyRequest,
     ) -> tag_20180828_models.CreatePolicyResponse:
         """
+        ###
         This topic provides an example on how to call the API operation to create a tag policy named `test`. In this example, the Tag Policy feature in multi-account mode is used. The tag policy defines that resources to which the `CostCenter:Beijing` or `CostCenter:Shanghai` tag is added are compliant and other resources are not compliant.
         
         @param request: CreatePolicyRequest
@@ -514,6 +518,7 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> tag_20180828_models.CreateTagsResponse:
         """
+        ###
         A preset tag is a tag that you create in advance and is available for the resources in all regions. You can create preset tags in the stage of tag planning and add them to specific resources in the stage of tag implementation. When you create a preset tag, you can specify only the tag key. You can specify a tag value in the future.
         This topic provides an example on how to call the API operation to create a preset tag whose tag key is `Environment` to indicate the business environment.
         
@@ -558,6 +563,7 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> tag_20180828_models.CreateTagsResponse:
         """
+        ###
         A preset tag is a tag that you create in advance and is available for the resources in all regions. You can create preset tags in the stage of tag planning and add them to specific resources in the stage of tag implementation. When you create a preset tag, you can specify only the tag key. You can specify a tag value in the future.
         This topic provides an example on how to call the API operation to create a preset tag whose tag key is `Environment` to indicate the business environment.
         
@@ -601,6 +607,7 @@ class Client(OpenApiClient):
         request: tag_20180828_models.CreateTagsRequest,
     ) -> tag_20180828_models.CreateTagsResponse:
         """
+        ###
         A preset tag is a tag that you create in advance and is available for the resources in all regions. You can create preset tags in the stage of tag planning and add them to specific resources in the stage of tag implementation. When you create a preset tag, you can specify only the tag key. You can specify a tag value in the future.
         This topic provides an example on how to call the API operation to create a preset tag whose tag key is `Environment` to indicate the business environment.
         
@@ -615,6 +622,7 @@ class Client(OpenApiClient):
         request: tag_20180828_models.CreateTagsRequest,
     ) -> tag_20180828_models.CreateTagsResponse:
         """
+        ###
         A preset tag is a tag that you create in advance and is available for the resources in all regions. You can create preset tags in the stage of tag planning and add them to specific resources in the stage of tag implementation. When you create a preset tag, you can specify only the tag key. You can specify a tag value in the future.
         This topic provides an example on how to call the API operation to create a preset tag whose tag key is `Environment` to indicate the business environment.
         
@@ -1077,6 +1085,8 @@ class Client(OpenApiClient):
     ) -> tag_20180828_models.DisablePolicyTypeResponse:
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.open_type):
+            query['OpenType'] = request.open_type
         if not UtilClient.is_unset(request.owner_account):
             query['OwnerAccount'] = request.owner_account
         if not UtilClient.is_unset(request.owner_id):
@@ -1115,6 +1125,8 @@ class Client(OpenApiClient):
     ) -> tag_20180828_models.DisablePolicyTypeResponse:
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.open_type):
+            query['OpenType'] = request.open_type
         if not UtilClient.is_unset(request.owner_account):
             query['OwnerAccount'] = request.owner_account
         if not UtilClient.is_unset(request.owner_id):
@@ -1167,6 +1179,8 @@ class Client(OpenApiClient):
     ) -> tag_20180828_models.EnablePolicyTypeResponse:
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.open_type):
+            query['OpenType'] = request.open_type
         if not UtilClient.is_unset(request.owner_account):
             query['OwnerAccount'] = request.owner_account
         if not UtilClient.is_unset(request.owner_id):
@@ -1205,6 +1219,8 @@ class Client(OpenApiClient):
     ) -> tag_20180828_models.EnablePolicyTypeResponse:
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.open_type):
+            query['OpenType'] = request.open_type
         if not UtilClient.is_unset(request.owner_account):
             query['OwnerAccount'] = request.owner_account
         if not UtilClient.is_unset(request.owner_id):
@@ -1748,6 +1764,8 @@ class Client(OpenApiClient):
         """
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.open_type):
+            query['OpenType'] = request.open_type
         if not UtilClient.is_unset(request.owner_account):
             query['OwnerAccount'] = request.owner_account
         if not UtilClient.is_unset(request.owner_id):
@@ -1793,6 +1811,8 @@ class Client(OpenApiClient):
         """
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.open_type):
+            query['OpenType'] = request.open_type
         if not UtilClient.is_unset(request.owner_account):
             query['OwnerAccount'] = request.owner_account
         if not UtilClient.is_unset(request.owner_id):
@@ -2392,7 +2412,10 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> tag_20180828_models.ListSupportResourceTypesResponse:
         """
-        This topic provides an example on how to call the API operation to query the resource types supported by tags.
+        ### [](#)Call examples
+        *   Query a list of resource types supported by TagResources or UntagResources. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22,%22SupportCode%22:%22TAG_CONSOLE_SUPPORT%22%7D).
+        *   Query a list of resource types supported by ListTagResources or ListResourcesByTag. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22%7D).
+        *   Query a list of resource types that support createdby tags. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22,%22SupportCode%22:%22CREATED_BY_TAG_CONSOLE_SUPPORT%22%7D).
         
         @param request: ListSupportResourceTypesRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -2445,7 +2468,10 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> tag_20180828_models.ListSupportResourceTypesResponse:
         """
-        This topic provides an example on how to call the API operation to query the resource types supported by tags.
+        ### [](#)Call examples
+        *   Query a list of resource types supported by TagResources or UntagResources. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22,%22SupportCode%22:%22TAG_CONSOLE_SUPPORT%22%7D).
+        *   Query a list of resource types supported by ListTagResources or ListResourcesByTag. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22%7D).
+        *   Query a list of resource types that support createdby tags. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22,%22SupportCode%22:%22CREATED_BY_TAG_CONSOLE_SUPPORT%22%7D).
         
         @param request: ListSupportResourceTypesRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -2497,7 +2523,10 @@ class Client(OpenApiClient):
         request: tag_20180828_models.ListSupportResourceTypesRequest,
     ) -> tag_20180828_models.ListSupportResourceTypesResponse:
         """
-        This topic provides an example on how to call the API operation to query the resource types supported by tags.
+        ### [](#)Call examples
+        *   Query a list of resource types supported by TagResources or UntagResources. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22,%22SupportCode%22:%22TAG_CONSOLE_SUPPORT%22%7D).
+        *   Query a list of resource types supported by ListTagResources or ListResourcesByTag. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22%7D).
+        *   Query a list of resource types that support createdby tags. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22,%22SupportCode%22:%22CREATED_BY_TAG_CONSOLE_SUPPORT%22%7D).
         
         @param request: ListSupportResourceTypesRequest
         @return: ListSupportResourceTypesResponse
@@ -2510,7 +2539,10 @@ class Client(OpenApiClient):
         request: tag_20180828_models.ListSupportResourceTypesRequest,
     ) -> tag_20180828_models.ListSupportResourceTypesResponse:
         """
-        This topic provides an example on how to call the API operation to query the resource types supported by tags.
+        ### [](#)Call examples
+        *   Query a list of resource types supported by TagResources or UntagResources. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22,%22SupportCode%22:%22TAG_CONSOLE_SUPPORT%22%7D).
+        *   Query a list of resource types supported by ListTagResources or ListResourcesByTag. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22%7D).
+        *   Query a list of resource types that support createdby tags. For more information, see [Example](https://api.alibabacloud.com/api/Tag/2018-08-28/ListSupportResourceTypes?tab=DEBUG\\&params=%7B%22RegionId%22:%22cn-hangzhou%22,%22SupportCode%22:%22CREATED_BY_TAG_CONSOLE_SUPPORT%22%7D).
         
         @param request: ListSupportResourceTypesRequest
         @return: ListSupportResourceTypesResponse